@@ -7,7 +7,7 @@ namespace WoWCombatLogParser.Common.Events
     [DebuggerDisplay("{Timestamp} {Event} {Source} {Destination}")]
     public partial class CompoundEventSection : EventSection
     {
-        public Unit Source { get; } = new Unit();
-        public Unit Destination { get; } = new Unit();
+        public Unit Source { get; set; } = new();
+        public Unit Destination { get; set; } = new();
     }
 }
