@@ -5,6 +5,6 @@ namespace WoWCombatLogParser.Common.Events
 {
     public abstract class PrefixAbilityBase : EventSection
     {
-        public Ability Spell { get; } = new Ability();
+        public Ability Spell { get; set; } = new();
     }
 }
