@@ -5,6 +5,6 @@ namespace WoWCombatLogParser.Common.Events
 {
     public abstract class SuffixAbilityBase : EventSection
     {
-        public Ability ExtraSpell { get; } = new Ability();
+        public Ability ExtraSpell { get; set; } = new();
     }
 }
